+++ src/xurrent/core.py
@@ -1,4 +1,6 @@
 from __future__ import annotations  # Needed for forward references
+from datetime import datetime
+import time
 import requests
 import logging
 import json
@@ -17,6 +19,11 @@ class JsonSerializableDict(dict):
                 # Recursively call to_dict on nested JsonSerializableDict objects
                 if isinstance(value, JsonSerializableDict):
                     result[key] = value.to_dict()
+                elif isinstance(value, list):
+                    #call to_dict on each item in the list
+                    result[key] = [item.to_dict() for item in value]
+                elif isinstance(value, datetime):
+                    result[key] = value.isoformat()
                 else:
                     result[key] = value
         return result
@@ -28,6 +35,7 @@ class JsonSerializableDict(dict):
 
 class XurrentApiHelper:
     api_user: Person # Forward declaration with a string
+    api_user_teams: List[Team] # Forward declaration with a string
 
     def __init__(self, base_url, api_key, api_account, resolve_user=True):
         self.base_url = base_url
@@ -38,6 +46,7 @@ class XurrentApiHelper:
             # Import Person lazily
             from .people import Person
             self.api_user = Person.get_me(self)
+            self.api_user_teams = self.api_user.get_teams()
 
     def __append_per_page(self, uri, per_page=100):
         """