+++ ChangeLog.md
@@ -1,5 +1,30 @@
 # Change Log
 
+## v0.0.2.9
+
+### New Features
+
+- Request, Workflow, Task, Person, Team: add non static methods: ref_str() --> return a reference string
+- Request: add RequestCategory enum
+- core: JSONSerializableDict: handle datetime and list of objects
+- Workflow: add WorkflowCategory enum
+- Workflow: use WorkflowCategory and WorkflowStatus enums on instantiation
+- Team: add Team class
+- Team: add enum TeamPredefinedFilter
+- People: add non static methods: get_teams
+- Tests: add tests for Request
+- Tests: add pre-commit hooks yaml file
+
+### Bugfixes
+
+- Person, Workflow, Task: inherit JsonSerializableDict --> make serializable
+- Request: close: make it possible to close a without a note (using default note)
+
+### Breaking Changes
+
+- Request: request.created_by, request.requested_by, request.requested_for, request.member are now Person objects
+- Workflow: workflow.manager is now a Person object
+
 ## v0.0.2.8
 
 ### Bug Fixes